chore(webapp): consolidate auth utils
closes #902
