+++ measure-web-app/app/[teamId]/layout.tsx
@@ -6,8 +6,7 @@ import { usePathname } from 'next/navigation';
 import { useRouter } from 'next/navigation';
 import TeamSwitcher, { TeamsSwitcherStatus } from "../components/team_switcher";
 import { TeamsApiStatus, emptyTeam, fetchTeamsFromServer } from "../api/api_calls";
-import { logout } from "../utils/auth_utils";
-import { auth } from "@/utils/auth";
+import { auth, logout } from "@/app/utils/auth/auth";
 
 export default function DashboardLayout({
   children,
